feat: Contacts Active Status
* safety actions active and inactive state makred
